fix: supabase 미들웨어 에러 관련 수정
