+++ src/middleware.ts
@@ -1,50 +1,20 @@
-import { createServerClient } from '@supabase/ssr';
-import { type NextRequest, NextResponse } from 'next/server';
+import { NextResponse } from 'next/server';
+import type { NextRequest } from 'next/server';
+import { updateSession } from './utils/supabase/middleware';
+import { createClient, getIsLogin } from './utils/supabase/server';
 
-export const updateSession = async (request: NextRequest) => {
-  // This `try/catch` block is only here for the interactive tutorial.
-  // Feel free to remove once you have Supabase connected.
-  try {
-    // Create an unmodified response
-    let response = NextResponse.next({
-      request: {
-        headers: request.headers
-      }
-    });
+export async function middleware(request: NextRequest) {
+  if (request.nextUrl.pathname.includes('/cart')) {
+    const isLogin = await getIsLogin();
 
-    const supabase = createServerClient(
-      process.env.NEXT_PUBLIC_SUPABASE_URL!,
-      process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-      {
-        cookies: {
-          getAll() {
-            return request.cookies.getAll();
-          },
-          setAll(cookiesToSet) {
-            cookiesToSet.forEach(({ name, value }) => request.cookies.set(name, value));
-            response = NextResponse.next({
-              request
-            });
-            cookiesToSet.forEach(({ name, value, options }) => response.cookies.set(name, value, options));
-          }
-        }
-      }
-    );
-
-    // This will refresh session if expired - required for Server Components
-    // https://supabase.com/docs/guides/auth/server-side/nextjs
-    await supabase.auth.getUser();
-
-    return response;
-  } catch (e) {
-    console.log(e);
-    // If you are here, a Supabase client could not be created!
-    // This is likely because you have not set up environment variables.
-    // Check out http://localhost:3000 for Next Steps.
-    return NextResponse.next({
-      request: {
-        headers: request.headers
-      }
-    });
+    if (!isLogin) {
+      return NextResponse.redirect(new URL('/sign-in', request.url));
+    }
   }
+
+  await updateSession(request);
+}
+
+export const config = {
+  matcher: '/((?!api|_next/static|_next/image|favicon.ico).*)'
 };
